requirements.txt
Django==1.3
recaptcha-client==1.0.6
